+++ example/by_address/by_address.cpp
@@ -59,12 +59,39 @@ motor_status_str(Sesame::motor_status_t status) {
 // Sesameからの通知がある毎に呼び出される(変化がある場合のみ通知されている模様)
 void
 status_update(SesameClient& client, SesameClient::Status status) {
-	if (status != last_status) {
+	static bool received = false;
+	if (!received || status != last_status) {
 		Serial.printf("Status in_lock=%u,in_unlock=%u,tgt=%d,pos=%d,volt=%.2f,batt_pct=%.2f,batt_crit=%u,motor_status=%s\n",
 		              status.in_lock(), status.in_unlock(), status.target(), status.position(), status.voltage(), status.battery_pct(),
 		              status.battery_critical(), motor_status_str(status.motor_status()));
+		if (!received ||
+		    (client.get_model() == Sesame::model_t::sesame_bot && status.motor_status() == Sesame::motor_status_t::idle &&
+		     last_status.motor_status() != Sesame::motor_status_t::idle) ||
+		    (client.get_model() != Sesame::model_t::sesame_bot && status.in_lock() != last_status.in_lock())) {
+			// 最新の操作履歴を読み出し可能
+			// 読み出し内容はコールバックで取得する
+			// ここでは初回の状態受信時と、施錠状態の変化時にrequest_history()を呼び出している
+			client.request_history();
+		}
 		last_status = status;
+		received = true;
 	}
+}
+
+// 履歴取得コールバック
+// request_history()を実行するとコールバックされる
+// SESAMEから失敗応答があった場合は history.type = Sesame::history_type_t::none となる
+// 応答がない場合等、呼び出されない可能性もある
+void
+receive_history(SesameClient& client, const SesameClient::History& history) {
+	if (history.type == Sesame::history_type_t::none) {
+		Serial.println("Empty history");
+		return;
+	}
+	struct tm tm;
+	gmtime_r(&history.time, &tm);
+	Serial.printf("History type=%u, %04d/%02d/%02d %02d:%02d:%02d, tag(%u)=%s\n", static_cast<uint8_t>(history.type),
+	              tm.tm_year + 1900, tm.tm_mon + 1, tm.tm_mday, tm.tm_hour, tm.tm_min, tm.tm_sec, history.tag_len, history.tag);
 }
 
 void
@@ -93,6 +120,8 @@ setup() {
 	client.set_state_callback([](auto& client, auto state) { sesame_state = state; });
 	// Sesame状態コールバックを設定
 	client.set_status_callback(status_update);
+	// 履歴受信コールバックを設定
+	client.set_history_callback(receive_history);
 }
 
 enum class app_state {
@@ -163,7 +192,7 @@ loop() {
 					state = app_state::pre_lock;
 				}
 			} else {
-				// 接続後、切断されるとSesameClientの状態がidleになる
+				// 接続後、認証失敗で切断されるとSesameClientの状態がidleになる
 				if (client.get_state() == SesameClient::state_t::idle) {
 					Serial.println("Failed to authenticate");
 					state = app_state::pre_close;
@@ -178,7 +207,7 @@ loop() {
 				}
 				last_operated = millis();
 				if (client.get_model() == Sesame::model_t::sesame_bot) {
-					// セサミbotは click() API を利用可能
+					// セサミbotは lock() unlock() に加えて click() API を利用可能
 					// click() はセサミアプリでボタンをタップした場合と同じ動作
 					state = app_state::pre_click;
 				} else {
@@ -195,7 +224,7 @@ loop() {
 					Serial.println("Failed to send click command");
 				}
 				last_operated = millis();
-				state = app_state::pre_close;
+				state = app_state::idle;
 			}
 			break;
 		case app_state::pre_close: