+++ test/test_main.cpp
@@ -20,7 +20,7 @@
 #endif
 
 #define TEST_UTILITY 1
-#define TEST_BLE 1
+#define TEST_BLE 0
 
 namespace util = libsesame3bt::util;
 using libsesame3bt::Sesame;
@@ -53,22 +53,54 @@ test_truncate_utf8() {
 }
 
 void
-test_vol_pct() {
-	TEST_ASSERT_EQUAL_FLOAT(0.0f, SesameClient::Status::voltage_to_pct(0.0f));
-	TEST_ASSERT_EQUAL_FLOAT(0.0f, SesameClient::Status::voltage_to_pct(4.6f));
-	TEST_ASSERT_EQUAL_FLOAT(100.0f, SesameClient::Status::voltage_to_pct(6.0f));
-	TEST_ASSERT_EQUAL_FLOAT(100.0f, SesameClient::Status::voltage_to_pct(7.0f));
-	TEST_ASSERT_EQUAL_FLOAT(50.0f, SesameClient::Status::voltage_to_pct(5.8f));
-	TEST_ASSERT_EQUAL_FLOAT(40.0f, SesameClient::Status::voltage_to_pct(5.7f));
-	TEST_ASSERT_EQUAL_FLOAT(32.0f, SesameClient::Status::voltage_to_pct(5.6f));
-	TEST_ASSERT_EQUAL_FLOAT(21.0f, SesameClient::Status::voltage_to_pct(5.4f));
-	TEST_ASSERT_EQUAL_FLOAT(13.0f, SesameClient::Status::voltage_to_pct(5.2f));
-	TEST_ASSERT_EQUAL_FLOAT(10.0f, SesameClient::Status::voltage_to_pct(5.1f));
-	TEST_ASSERT_EQUAL_FLOAT(7.0f, SesameClient::Status::voltage_to_pct(5.0f));
-	TEST_ASSERT_EQUAL_FLOAT(3.0f, SesameClient::Status::voltage_to_pct(4.8f));
+test_cleanup_tail_utf8() {
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("", 0));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\0", 1));
+	TEST_ASSERT_EQUAL(1, util::cleanup_tail_utf8("A"));
+	TEST_ASSERT_EQUAL(2, util::cleanup_tail_utf8(u8"Ā"));
+	TEST_ASSERT_EQUAL(3, util::cleanup_tail_utf8(u8"あ"));
+	TEST_ASSERT_EQUAL(5, util::cleanup_tail_utf8(u8"AあB"));
+	TEST_ASSERT_EQUAL(4, util::cleanup_tail_utf8(u8"😀"));
 
-	float upper = SesameClient::Status::voltage_to_pct(5.83f);
-	float lower = SesameClient::Status::voltage_to_pct(5.82f);
+	// incorrect 2 bytes code
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xc4"));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xc4\x00", 2));
+
+	// incorrect 3 bytes code
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xe0\x81"));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xe0"));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xe0\x81\x00", 3));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xe0\x20\x00", 3));
+
+	// incorrect 4 bytes code
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xf0\x9f\x98"));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xf0\x9f"));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xf0"));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xf0\x9f\x98\x00", 4));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xf0\x30\x98\x80", 4));
+	TEST_ASSERT_EQUAL(0, util::cleanup_tail_utf8("\xf0\x9f\x40\x80", 4));
+
+	// mixture
+	TEST_ASSERT_EQUAL(9, util::cleanup_tail_utf8("あ😀Ā\xe0\x81\x00", 12));
+}
+
+void
+test_vol_pct() {
+	TEST_ASSERT_EQUAL_FLOAT(0.0f, SesameClient::Status::voltage_to_pct(0.00f));
+	TEST_ASSERT_EQUAL_FLOAT(0.0f, SesameClient::Status::voltage_to_pct(4.60f));
+	TEST_ASSERT_EQUAL_FLOAT(100.0f, SesameClient::Status::voltage_to_pct(5.85f));
+	TEST_ASSERT_EQUAL_FLOAT(100.0f, SesameClient::Status::voltage_to_pct(6.00f));
+	TEST_ASSERT_EQUAL_FLOAT(50.0f, SesameClient::Status::voltage_to_pct(5.60f));
+	TEST_ASSERT_EQUAL_FLOAT(40.0f, SesameClient::Status::voltage_to_pct(5.55f));
+	TEST_ASSERT_EQUAL_FLOAT(32.0f, SesameClient::Status::voltage_to_pct(5.50f));
+	TEST_ASSERT_EQUAL_FLOAT(21.0f, SesameClient::Status::voltage_to_pct(5.40f));
+	TEST_ASSERT_EQUAL_FLOAT(13.0f, SesameClient::Status::voltage_to_pct(5.20f));
+	TEST_ASSERT_EQUAL_FLOAT(10.0f, SesameClient::Status::voltage_to_pct(5.10f));
+	TEST_ASSERT_EQUAL_FLOAT(7.0f, SesameClient::Status::voltage_to_pct(5.00f));
+	TEST_ASSERT_EQUAL_FLOAT(3.0f, SesameClient::Status::voltage_to_pct(4.80f));
+
+	float upper = SesameClient::Status::voltage_to_pct(5.84f);
+	float lower = SesameClient::Status::voltage_to_pct(5.61f);
 	TEST_ASSERT_MESSAGE(upper > lower, "upper > lower");
 	TEST_ASSERT_MESSAGE(upper < 100.0f, "upper < 100%");
 	TEST_ASSERT_MESSAGE(lower > 50.0f, "lower > 50%");
@@ -122,6 +154,7 @@ setup() {
 	UNITY_BEGIN();
 #if TEST_UTILITY
 	RUN_TEST(test_truncate_utf8);
+	RUN_TEST(test_cleanup_tail_utf8);
 	RUN_TEST(test_vol_pct);
 #endif
 #if TEST_BLE